Move AccessPolicy to model_editor
