+++ openstudiocore/src/energyplus/ForwardTranslator/ForwardTranslateVersion.cpp
@@ -31,6 +31,7 @@
 #include "../../model/Version.hpp"
 #include <utilities/idd/Version_FieldEnums.hxx>
 #include "../../utilities/idd/IddEnums.hpp"
+#include "../../utilities/idd/IddObject.hpp"
 #include <utilities/idd/IddEnums.hxx>
 
 #include "../../utilities/core/Assert.hpp"
